Move code to src
This cleans up the root folder. There are already a lot of others files
like package management files or configuration files. This makes it less
convoluted.
