+++ deno.json
@@ -1,7 +1,7 @@
 {
   "tasks": {
-    "dev": "deno run  --allow-net --allow-env --env-file dev.js",
-    "start": "deno run  --allow-net --allow-env --env-file main.js",
+    "dev": "deno run  --allow-net --allow-env --env-file src/dev.js",
+    "start": "deno run  --allow-net --allow-env --env-file src/main.js",
     "lint:fix": "deno lint --fix"
   },
   "lint": {
